refactor: create trestle trash util module
